requirements.txt
coverage==4.4.1
Django==1.11.3
django-bootstrap3==9.0.0
django-braces==1.11.0
django-model-utils==3.0.0
django-notifications-hq==1.3
humanize==0.5.1
jsonfield==2.0.2
markdown2==2.3.4
olefile==0.44
Pillow==4.3.0
pytz==2017.2
